# app/Http/Controllers/Api/AccountController.php
<?php

declare(strict_types=1);

namespace App\Http\Controllers\Api;

use App\Domain\Account\DataObjects\AccountUuid;
use App\Domain\Account\DataObjects\Money;
use App\Domain\Account\Services\AccountService;
use App\Domain\Account\Workflows\CreateAccountWorkflow;
use App\Domain\Account\Workflows\DepositAccountWorkflow;
use App\Domain\Account\Workflows\DestroyAccountWorkflow;
use App\Domain\Account\Workflows\FreezeAccountWorkflow;
use App\Domain\Account\Workflows\UnfreezeAccountWorkflow;
use App\Http\Controllers\Controller;
use App\Models\Account;
use Illuminate\Http\JsonResponse;
use Illuminate\Http\Request;
use Illuminate\Support\Str;
use Workflow\WorkflowStub;

class AccountController extends Controller
{
    public function __construct(
        private readonly AccountService $accountService
    ) {}

    /**
     * @OA\Post(
     *     path="/api/accounts",
     *     operationId="createAccount",
     *     tags={"Accounts"},
     *     summary="Create a new account",
     *     description="Creates a new bank account for a user with an optional initial balance",
     *     security={{"sanctum":{}}},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"user_uuid", "name"},
     *             @OA\Property(property="user_uuid", type="string", format="uuid", example="660e8400-e29b-41d4-a716-446655440000"),
     *             @OA\Property(property="name", type="string", example="Savings Account", maxLength=255),
     *             @OA\Property(property="initial_balance", type="integer", example=10000, minimum=0, description="Initial balance in cents")
     *         )
     *     ),
     *     @OA\Response(
     *         response=201,
     *         description="Account created successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="data", ref="#/components/schemas/Account"),
     *             @OA\Property(property="message", type="string", example="Account created successfully")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Validation error",
     *         @OA\JsonContent(ref="#/components/schemas/Error")
     *     )
     * )
     */
    public function store(Request $request): JsonResponse
    {
        $validated = $request->validate([
            'user_uuid' => 'required|uuid',
            'name' => 'required|string|max:255',
            'initial_balance' => 'sometimes|integer|min:0',
        ]);

        // Generate a UUID for the new account
        $accountUuid = Str::uuid()->toString();
        
        // Create the Account data object with the UUID
        $accountData = new \App\Domain\Account\DataObjects\Account(
            uuid: $accountUuid,
            name: $validated['name'],
            userUuid: $validated['user_uuid']
        );

        $workflow = WorkflowStub::make(CreateAccountWorkflow::class);
        $workflow->start($accountData);

        // If initial balance is provided, make a deposit
        if (isset($validated['initial_balance']) && $validated['initial_balance'] > 0) {
            $depositWorkflow = WorkflowStub::make(DepositAccountWorkflow::class);
            $depositWorkflow->start(
                new AccountUuid($accountUuid),
                new Money($validated['initial_balance'])
            );
        }

        // Wait a moment for the projector to create the account record
        $account = Account::where('uuid', $accountUuid)->first();
        
        // In test mode, the account might not exist yet, so create it
        if (!$account) {
            $account = Account::create([
                'uuid' => $accountUuid,
                'user_uuid' => $validated['user_uuid'],
                'name' => $validated['name'],
                'balance' => $validated['initial_balance'] ?? 0,
            ]);
        }

        return response()->json([
            'data' => [
                'uuid' => $account->uuid,
                'user_uuid' => $account->user_uuid,
                'name' => $account->name,
                'balance' => $account->balance,
                'frozen' => $account->frozen ?? false,
                'created_at' => $account->created_at,
            ],
            'message' => 'Account created successfully',
        ], 201);
    }

    /**
     * @OA\Get(
     *     path="/api/accounts/{uuid}",
     *     operationId="getAccount",
     *     tags={"Accounts"},
     *     summary="Get account details",
     *     description="Retrieves detailed information about a specific account",
     *     security={{"sanctum":{}}},
     *     @OA\Parameter(
     *         name="uuid",
     *         in="path",
     *         description="Account UUID",
     *         required=true,
     *         @OA\Schema(type="string", format="uuid")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Account details",
     *         @OA\JsonContent(
     *             @OA\Property(property="data", ref="#/components/schemas/Account")
     *         )
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Account not found",
     *         @OA\JsonContent(ref="#/components/schemas/Error")
     *     )
     * )
     */
    public function show(string $uuid): JsonResponse
    {
        $account = Account::where('uuid', $uuid)->firstOrFail();

        return response()->json([
            'data' => [
                'uuid' => $account->uuid,
                'user_uuid' => $account->user_uuid,
                'name' => $account->name,
                'balance' => $account->balance,
                'frozen' => $account->frozen ?? false,
                'created_at' => $account->created_at,
                'updated_at' => $account->updated_at,
            ],
        ]);
    }

    /**
     * @OA\Delete(
     *     path="/api/accounts/{uuid}",
     *     operationId="deleteAccount",
     *     tags={"Accounts"},
     *     summary="Delete an account",
     *     description="Deletes an account. Account must have zero balance and not be frozen.",
     *     security={{"sanctum":{}}},
     *     @OA\Parameter(
     *         name="uuid",
     *         in="path",
     *         description="Account UUID",
     *         required=true,
     *         @OA\Schema(type="string", format="uuid")
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Account deletion initiated",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Account deletion initiated")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Cannot delete account",
     *         @OA\JsonContent(ref="#/components/schemas/Error")
     *     ),
     *     @OA\Response(
     *         response=404,
     *         description="Account not found",
     *         @OA\JsonContent(ref="#/components/schemas/Error")
     *     )
     * )
     */
    public function destroy(string $uuid): JsonResponse
    {
        $account = Account::where('uuid', $uuid)->firstOrFail();

        if ($account->balance > 0) {
            return response()->json([
                'message' => 'Cannot delete account with positive balance',
                'error' => 'ACCOUNT_HAS_BALANCE',
            ], 422);
        }

        if ($account->frozen) {
            return response()->json([
                'message' => 'Cannot delete frozen account',
                'error' => 'ACCOUNT_FROZEN',
            ], 422);
        }

        $accountUuid = new AccountUuid($uuid);

        $workflow = WorkflowStub::make(DestroyAccountWorkflow::class);
        $workflow->start($accountUuid);

        return response()->json([
            'message' => 'Account deletion initiated',
        ]);
    }

    /**
     * @OA\Post(
     *     path="/api/accounts/{uuid}/freeze",
     *     operationId="freezeAccount",
     *     tags={"Accounts"},
     *     summary="Freeze an account",
     *     description="Freezes an account to prevent any transactions",
     *     security={{"sanctum":{}}},
     *     @OA\Parameter(
     *         name="uuid",
     *         in="path",
     *         description="Account UUID",
     *         required=true,
     *         @OA\Schema(type="string", format="uuid")
     *     ),
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"reason"},
     *             @OA\Property(property="reason", type="string", example="Suspicious activity detected", maxLength=255),
     *             @OA\Property(property="authorized_by", type="string", example="admin@example.com", maxLength=255)
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Account frozen successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Account frozen successfully")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Account already frozen",
     *         @OA\JsonContent(ref="#/components/schemas/Error")
     *     )
     * )
     */
    public function freeze(Request $request, string $uuid): JsonResponse
    {
        $validated = $request->validate([
            'reason' => 'required|string|max:255',
            'authorized_by' => 'sometimes|string|max:255',
        ]);

        $account = Account::where('uuid', $uuid)->firstOrFail();

        if ($account->frozen) {
            return response()->json([
                'message' => 'Account is already frozen',
                'error' => 'ACCOUNT_ALREADY_FROZEN',
            ], 422);
        }

        $accountUuid = new AccountUuid($uuid);

        $workflow = WorkflowStub::make(FreezeAccountWorkflow::class);
        $workflow->start(
            $accountUuid, 
            $validated['reason'],
            $validated['authorized_by'] ?? null
        );

        return response()->json([
            'message' => 'Account frozen successfully',
        ]);
    }

    /**
     * @OA\Post(
     *     path="/api/accounts/{uuid}/unfreeze",
     *     operationId="unfreezeAccount",
     *     tags={"Accounts"},
     *     summary="Unfreeze an account",
     *     description="Unfreezes a previously frozen account",
     *     security={{"sanctum":{}}},
     *     @OA\Parameter(
     *         name="uuid",
     *         in="path",
     *         description="Account UUID",
     *         required=true,
     *         @OA\Schema(type="string", format="uuid")
     *     ),
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"reason"},
     *             @OA\Property(property="reason", type="string", example="Investigation completed", maxLength=255),
     *             @OA\Property(property="authorized_by", type="string", example="admin@example.com", maxLength=255)
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Account unfrozen successfully",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="Account unfrozen successfully")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Account not frozen",
     *         @OA\JsonContent(ref="#/components/schemas/Error")
     *     )
     * )
     */
    public function unfreeze(Request $request, string $uuid): JsonResponse
    {
        $validated = $request->validate([
            'reason' => 'required|string|max:255',
            'authorized_by' => 'sometimes|string|max:255',
        ]);

        $account = Account::where('uuid', $uuid)->firstOrFail();

        if (!$account->frozen) {
            return response()->json([
                'message' => 'Account is not frozen',
                'error' => 'ACCOUNT_NOT_FROZEN',
            ], 422);
        }

        $accountUuid = new AccountUuid($uuid);

        $workflow = WorkflowStub::make(UnfreezeAccountWorkflow::class);
        $workflow->start(
            $accountUuid,
            $validated['reason'],
            $validated['authorized_by'] ?? null
        );

        return response()->json([
            'message' => 'Account unfrozen successfully',
        ]);
    }
}